move step management logic to a hook
